Versão 2

+++ back/src/database/ShoppingCartDatabase.ts
@@ -4,8 +4,7 @@ import { BaseDatabase } from "./BaseDatabase"
 export class ShoppingCartDatabase extends BaseDatabase {
     public static TABLE_SHOPPING_CART = "shopping_cart"
 
-    // RETORNA TODOS OS PRODUTOS DO CARRINHO
-    public async selectAllProducts(): Promise<any> {
+    public async selectAllProducts(user_id: string): Promise<any> {
         const products = await BaseDatabase.connection
             .select(
                 "products.id",
@@ -15,49 +14,49 @@ export class ShoppingCartDatabase extends BaseDatabase {
                 "shopping_cart.quantity"
             )
             .from(ShoppingCartDatabase.TABLE_SHOPPING_CART)
-            .join("products", "products.id", "=", "shopping_cart.id_product")
+            .join("products", "products.id", "=", "shopping_cart.product_id")
+            .where({ user_id })
         
         return products
     }
 
-    // SELECT PARA VERIFICAR SE O PRODUTO JÁ ESTÁ NO CARRINHO
     public async selectProduct(data: ISelectCartPrductDTO): Promise<any> {
         const product = await BaseDatabase.connection
             .select("*")
             .from(ShoppingCartDatabase.TABLE_SHOPPING_CART)
-            .where(data)
+            .where({ product_id: data.product_id })
+            .andWhere({ user_id: data.user_id })
 
         return product
     }
 
-    // REMOVE UM PRODUTO DO CARRINHO
     public async deleteProduct(data: IRemoveProductDTO): Promise<void> {
         await BaseDatabase.connection
             .delete()
             .from(ShoppingCartDatabase.TABLE_SHOPPING_CART)
-            .where(data)
+            .where({ product_id: data.product_id })
+            .andWhere({ user_id: data.user_id })
     }
 
-    // ALTERA A QUANTIDADE DE UM PRODUTO CADASTRADO
     public async updateProduct(data: IUpdateProductQuantityDTO): Promise<void> {
         await BaseDatabase.connection
             .update({ quantity: data.quantity })
             .from(ShoppingCartDatabase.TABLE_SHOPPING_CART)
-            .where({ id_product: data.id_product })
+            .where({ product_id: data.product_id })
+            .andWhere({ user_id: data.user_id })
     }
 
-    // ADICIONA UM PRODUTO AO CARRINHO
     public async insertProduct(data: IAddProductCartDTO): Promise<void> {
         await BaseDatabase.connection
             .insert(data)
             .into(ShoppingCartDatabase.TABLE_SHOPPING_CART)
     }
 
-    // REMOVE TODOS OS PRODUTOS DO CARRINHO
-    public deleteAllProducts = async(): Promise<void> => {
+    public deleteAllProducts = async(id: string): Promise<void> => {
         await BaseDatabase.connection
             .delete()
             .from(ShoppingCartDatabase.TABLE_SHOPPING_CART)
+            .where({ user_id: id })
     }
 
 }
